requirements.txt
discord.py
supabase
openai
python-dotenv
requests
